Split into modules and add sum_tree B+tree based browser
