+++ crates/volt/src/components.rs
@@ -1,0 +1,1 @@
+pub mod adjustable_input;
