+++ crates/volt/src/theme.rs
@@ -1,0 +1,68 @@
+use gpui::{Rgba, rgb, rgba};
+
+pub struct ThemeColors {
+    pub accent: Rgba,
+    pub navbar_background_gradient_top: Rgba,
+    pub navbar_background_gradient_bottom: Rgba,
+    pub navbar_outline: Rgba,
+    pub navbar_widget: Rgba,
+    pub notification_background: Rgba,
+    pub notification_border: Rgba,
+    pub central_background: Rgba,
+    pub browser: Rgba,
+    pub browser_outline: Rgba,
+    pub browser_selected_button_fg: Rgba,
+    pub browser_unselected_button_fg: Rgba,
+    pub browser_unselected_hover_button_fg: Rgba,
+    pub browser_invalid_name_bg: Rgba,
+    pub browser_unselected_button_fg_invalid: Rgba,
+    pub browser_unselected_hover_button_fg_invalid: Rgba,
+    pub browser_folder_text: Rgba,
+    pub browser_folder_hover_text: Rgba,
+    pub playlist_bar: Rgba,
+    pub playlist_beat: Rgba,
+    pub bg_text: Rgba,
+    pub command_palette: Rgba,
+    pub command_palette_border: Rgba,
+    pub command_palette_text: Rgba,
+    pub command_palette_placeholder_text: Rgba,
+    pub playhead: Rgba,
+    pub playhead_hover: Rgba,
+    pub preview: Rgba,
+    pub hover: Rgba,
+}
+
+#[expect(clippy::unreadable_literal, reason = "these are hex codes")]
+pub fn default() -> ThemeColors {
+    ThemeColors {
+        accent: rgb(0xb6afff),
+        navbar_background_gradient_top: rgb(0x1e2132),
+        navbar_background_gradient_bottom: rgb(0x171825),
+        navbar_outline: rgb(0x453f67),
+        navbar_widget: rgba(0x07081520),
+        notification_background: rgb(0x1d1b2b),
+        notification_border: rgb(0x3d3b4b),
+        central_background: rgb(0x171825),
+        browser: rgb(0x171825),
+        browser_outline: rgb(0x28243e),
+        browser_selected_button_fg: rgb(0xffcf7b),
+        browser_unselected_button_fg: rgb(0x646d88),
+        browser_unselected_hover_button_fg: rgb(0x8591b5),
+        browser_invalid_name_bg: rgba(0xff000010),
+        browser_unselected_button_fg_invalid: rgb(0xa46d88),
+        browser_unselected_hover_button_fg_invalid: rgb(0xf591b5),
+        browser_folder_text: rgb(0x928ea7),
+        browser_folder_hover_text: rgb(0xece9ff),
+        playlist_bar: rgb(0x4c495f),
+        playlist_beat: rgb(0x2e2b3f),
+        bg_text: rgb(0x646987),
+        command_palette: rgb(0x1d1b2b),
+        command_palette_border: rgb(0x3d3b4b),
+        command_palette_text: rgb(0x928ea7),
+        command_palette_placeholder_text: rgba(0x928ea740),
+        playhead: rgb(0xf00000),
+        playhead_hover: rgba(0xf000007f),
+        preview: rgb(0xf000f0),
+        hover: rgba(0xffffff20),
+    }
+}
