+++ crates/volt/src/views.rs
@@ -1,0 +1,2 @@
+pub mod browser;
+pub mod playlist;
